requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
PyJWT==2.8.0
python-dotenv==1.0.0
bcrypt==4.1.2
pytest==7.4.3
pytest-flask==1.3.0
python-dateutil==2.8.2
marshmallow==3.20.1
flask-cors==4.0.0
flasgger==0.9.7.1
